Implemented complete server side authorization as well

+++ controllers/review.js
@@ -1,6 +1,10 @@
 const Review = require('../models/review');
 
 module.exports.postReview = async(req, res) => {
+    if(!req.user) {
+        req.flash('failure', 'Access Denied, Login first');
+        return res.redirect('/auth/login');
+    }
     const {id} = req.params;
     const newReview = new Review(req.body.review);
     await newReview.save();
@@ -11,7 +15,11 @@ module.exports.postReview = async(req, res) => {
 
 module.exports.deleteReview = async(req, res) => {
     const {id} = req.params;
-    const reqReview = await Review.findById(id);
+    const reqReview = await Review.findById(id).populate('author');
+    if(!req.user._id.equals(reqReview.author._id)) {
+        req.flash('failure', 'Access Denied, Login first');
+        return res.redirect('/auth/login');
+    }
     const reqListingId = reqReview.listing; 
     await Review.findByIdAndDelete(id);
     req.flash('success', 'Your review was deleted successfully!')
